+++ library/src/pru/servo.c
@@ -8,13 +8,13 @@
 
 #include <stdio.h>
 #include <stdint.h>
-#include <math.h> //for lroundf
+#include <math.h> //for lround
 #include <rc/pru.h>
 #include <rc/gpio.h>
 #include <rc/servo.h>
 #include <rc/time.h>
 
-#define TOL		0.01	// acceptable tolerance on floating point bounds
+#define TOL		0.01	// acceptable tolerance on doubleing point bounds
 #define GPIO_POWER_PIN	2,16	//gpio2.16 P8.36
 #define SERVO_PRU_CH	1	// PRU1
 #define SERVO_PRU_FW	"am335x-pru1-rc-servo-fw"
@@ -168,7 +168,7 @@ int rc_servo_send_pulse_us(int ch, int us)
 }
 
 
-int rc_servo_send_pulse_normalized(int ch, float input)
+int rc_servo_send_pulse_normalized(int ch, double input)
 {
 	int us;
 	if(input<(-1.5-TOL) || input>(1.5+TOL)){
@@ -176,32 +176,32 @@ int rc_servo_send_pulse_normalized(int ch, float input)
 		return -1;
 	}
 	// normal range is from 900 to 2100 for 120 degree servos
-	us = 1500 + lroundf((input*600.0));
+	us = 1500 + lround((input*600.0));
 	return rc_servo_send_pulse_us(ch, us);
 }
 
 
 
-int rc_servo_send_esc_pulse_normalized(int ch, float input)
+int rc_servo_send_esc_pulse_normalized(int ch, double input)
 {
 	int us;
 	if(input<(-0.1-TOL) || input>(1.0+TOL)){
 		fprintf(stderr,"ERROR in rc_servo_send_esc_pulse_normalized, normalized input must be between -0.1 & 1.0\n");
 		return -1;
 	}
-	us = esc_min_us + lroundf((input*(esc_max_us-esc_min_us)));
+	us = esc_min_us + lround((input*(esc_max_us-esc_min_us)));
 	return rc_servo_send_pulse_us(ch, us);
 }
 
 
 
-int rc_servo_send_oneshot_pulse_normalized(int ch, float input)
+int rc_servo_send_oneshot_pulse_normalized(int ch, double input)
 {
 	int us;
 	if(input<(-0.1-TOL) || input>(1.0+TOL)){
 		fprintf(stderr,"ERROR in rc_servo_send_oneshot_pulse_normalized, normalized input must be between -0.1 & 1.0\n");
 		return -1;
 	}
-	us = 125 + lroundf(input*125.0);
+	us = 125 + lround(input*125.0);
 	return rc_servo_send_pulse_us(ch, us);
 }
